Add conditional replacement string function, $ifcondition(expr1,expr2)
Allow emitting different code based on some test determined by a language module,
typically in a typemap.  This is used to have errors occurring in directorout
typemaps either throw exceptions or do nothing much if the currect method has an
exception specification.

The usage is to put a string like $CONDTAG(EXPR1,EXPR2) in a typemap or
feature being emitted, and then in the language module, calling
DohReplaceConditional("$CONDTAG", DohString*, int val).  Any occurrences
of $CONDTAG(...) are replaced by the EXPR1 if val is not zero, or
EXPR2 if it is zero.  EXPR1 and EXPR2 are separated by a comma, but may
contain commas as long as they are inside nested ( ) or inside quoted string,
OR (if neither of those) explicitly escaped with a '\'.  See the testsuite.c
for examples.