Change typemaps and java.cxx, to raise exceptions if possible on errors
This fixes the existing implementation that fails to signal error
conditions properly.  When a c/c++ layer makes a call to a director
implementation, the out return value may not be convertible (e.g. java
null into a std::string or a wrapped c++ object) or other (unlikely)
errors may occur like allocation failures.  In these cases, current
implementations sets a pending java exception that will not be visible
to the C++ caller, and will not be signalled until execution reenters
the JVM in possibly only weakly related code.

The change is to throw Swig::DirectorException when these situations
arise, which will always immediately interrupt execution and signal an
error to caller.  When the method being called has an exception
specification, however, the code can not throw arbitrary exceptions,
so in that case, the old behavior is maintained - there is no way to
signal an error.