Update tests for changes in director exception propagation
The C++ wrappers of director implementations in target language (java)
may now throw a Swig::DirectorException for certain error conditions.
The use of "except" and "director:except" features for methods must be
consistent to handle these "unexpected" exceptions. Use the generic macros
defined in generic_director_exceptions.i in the tests and add coverage
to the tests for the unexpected conditions.